fix install rc.wifi

+++ build_slackware_rootfs.sh
@@ -157,6 +157,10 @@ EOF
 
 
 setting_wifi() {
+    if [[ ! -f "$CWD/bin/$BOARD_NAME/rc.wifi" ]]; then
+        return 0
+    fi
+
     message "" "setting" "wifi"
     install -m755 -D "$CWD/bin/$BOARD_NAME/rc.wifi" "$CWD/$BUILD/$SOURCE/$ROOTFS/etc/rc.d/rc.wifi"
